ssh: Enable property_test/ssh_eqc_client_server for PropEr
PropEr does not support the grouped statem-tests.  This commits
calls the grouped functions from the functions supported by PropEr.
Optimized with memoization.

Previously only EQC was supported, but the changes to support PROPER is not
just a wrapper. Since I don't have access to eqc I can't test the changes
and therefore eqc is disabeled.
However, with access to eqc it ought to be quite easy to re-enable eqc by
studying the diff.