[cdv] Handle multiple "Yc" lines refering to the same refc binary
Usually, sub binaries are implemented as special sub binary terms
("Ys" lines). But if a sub binary is sent to another process, sub
binaries will be eliminated and incorporated into the ProcBin ("Yc"
line), and when that happens there can be several "Yc" lines that
refer to the same underlying binary object. Make sure that that the
crash dump viewer can handle multiple "Yc" lines referring to the same
binary object.