Share the code for semantically equivalent blocks
Share code for semantically equivalent blocks referred to to by `br`
and `switch` instructions.

A similar optimization is done in `beam_jump`, but doing it here as
well is beneficial as it may enable other optimizations. Also, if
there are many semantically equivalent clauses, this optimization can
substanstially decrease compilation times.