Turn some gcbifs to ubifs on 64bit systems
BIFs erlang:size/1, erlang:byte_size/1 and erlang:map_size/1 on 64bit
systems would never trigger GC, so they can be implemented as regular
non-gc BIFs - they are slightly more efficient.
The change is done only in the VM and the loader - the compiler always
produces gc_bif instructions for those BIFs.

The fact that size of bianries fits into small ints on 64bits is already
leveraged in the new bs_get/set_position instructions. If byte size of
binaries will fit into a small, so will a size of a map since the
size indicates number of words instead of bytes.

Additionally:
  * change erlang:map_size/1 implementation not to allocate a
    bigint if the hashmap size fits in a small, which should be
    the common case.
  * change the instruction names to i_bif1 and i_bif1_body for
    consistency
  * support %if sequences in make_tables .tab files similar to
    files generated by beam_makeops