[socket-nif] Add "partial success" to sendmsg
The sendmsg function attempts to send *one message*. But
its possible for the underlying software to fail to send
the *entire* message. So, instead of retrying itself, as
send does, the sendmsg function will now instead return
with {ok, Remaining}, leaving it to the caller to decide
what to do.

OTP-14831