ssh: Fix ssh_options checking for ext_info
A bug for SSH_MSG_EXT_INFO was fixed both for client and server. Before that fix, wrong
option was read for the information sent to the peer.

This commit adapts the option checking so that the correct option now used is available
not only for servers but also for clients.