Fix broken assertion on monitor release
We sometimes bump the refc without messing with the table, which
means that we sometimes decrement it while in the table, causing
the old assertion to fail.

The property we want to check is that neither end of the monitor is
present in the table when the monitor is deleted.