beam_ssa_type: Speed up type analysis for huge functions
The type analysis pass (`beam_ssa_type`) keeps the type information
for all variables that are in scope. For huge functions, the
`join_types/2` function could get really slow when joining two
maps with thousands of variables in each.

Use a conservative approach to discard type information for
variables only used once by a `br` or `switch` in the same
block as the variable is defined in.

This approach reduces the runtime for the `beam_ssa_type` pass
from a few minutes down to few seconds for this module:

https://github.com/aggelgian/cuter/blob/master/src/cuter_binlib.erl

Rejected approach: Calculate liveness information for all variables
and discard type information for any variable that would not be
used again. That turned out to not work because some optimizations
would invalidate the liveness (in particular, substitutions could
lengthen the lifetime for a variable). Trying to update the
liveness information when doing the optimizations would be tricky.

Noticed-by: Kostis Sagonas