fix dirty scheduler sleepers list
Fix problem where only a few of the online schedulers are woken up when job arrives.
Address this issue by making sure the sleepers list for the dirty run-queue
only includes schedulers that are asleep (not suspened ones) and only once.
No loops between SSI (scheduler sleep information) structures.

This occurs when schedulers are suspened.
In the case decreasing changing dirty_cpu_schedulers_online.
The first created scheduler are woken up to cover for the suspened scheduler (if it just was woken up to perform some job).
But if it was in the sleepers list it will be re-added.
Since dcpu_sched_ix_wake will not remove the scheduler from the sleepers list.

And even the suspended scheduler may have been in the sleepers list.
Then when the suspended scheduler are taken online again it will be readded to the sleepers list when entering sleep.