Use microops for BIFs
This allows bif1/2/3 to share the main part of the code.
The price is that we always need to copy all three temporary registers
when error handling in bodies, but that should be infrequent.

Additionally it makes it a bit harder to read the disasembly since now
the arguments to BIFs are in the reverse order.