erts: Fix seq_trace:reset_trace dirty gc bug
When seq_trace:reset_trace could be called while a
process was doing a dirty GC. This triggered a race
where all signals was moved to the internal signal
queue during the GC which in turn caused the a heap
overrun problem.

This fix makes it so that the main and msgq lock are
taken before the clear. This will make sure that we
are allowed to do the clear.