[socket-nif|test] Add a "global" logger
Added a global logger that make it possible to log
from the slave nodes (with "ease").

Also "fixed" the test case that failed on "older"
linux (Ubuntu 14.04). For now we let it skip
instead (we should really check the OS version).

Also corrected a couple of (ping-pong) cases for which
the buffer adjustments did not work.

OTP-14831