ssh: Wait a bit for the docker client to finnish in ssh_compat_SUITE
The testing with a remote client in a docker and a local server could
show some hazards.  For instance instructing the client to sftp a
file to the server is not necessarily ready in the server's file system
when the common_test is scheduled.

To avoid that the test suite dooms a file operation to not have been
performed, simply sleep at increasing times for a while with testing
if it is done.  After a while give up and signal a fault.