Optimise ext table dumping
Folding over a list of ops in a commit with orddict:append/3 is an
O(n^2) algorithm. It is faster to fold over the list in reverse order
and prepend each op to the list of already collected ops.

A further optimisation is using a map instead of an orddict.

Also, since external storage engines with disc_only_copies semantics
only need to be dumped at startup, filter out these ops right when
building the map for all other dump types.