Add start_distribution to kernel environment
Sometimes you may want to start Erlang without any
of its distribution services. This commit adds an
environment configuration that allows so.

Because the servers we don't start here are a subset
of the servers not started on minimal mode, we do
have a guarantee that the system can still operate
as the system operates without those apps on minimal
mode.