Refactor string operands
There are two instructions that take string operands:

     {bs_put_string,Fail,NumberOfBytes,{string,String}}
     {bs_match_string,Fail,Register,NumberOfBits,{string,String}}

In the canonical BEAM code that is passed to beam_asm, string String
is currently represented as a list. (The string in bs_match_string is
a bitstring before the beam_z compiler pass.) That is wasteful,
because there will be unnecessary conversions between lists and
binaries.

Change the representation of String to be a binary.

Furthermore, bs_put_string is an optimization of a bs_put_binary
instruction with a literal binary operand. Currently, the
bs_put_string instruction is introduced in beam_bs. Delay the
introduction of bs_put_string to the beam_z pass. That will simplify
optimizations and allow us to do the optimization currently done
in beam_bs in a SSA pass in a future commit.