Add application:set_env/1 and application:set_env/2
It is equivalent to calling application:set_env/4 on
each application individially, except it is more efficient.

When given duplicate apps or duplicate keys, set_env/1
and set_env/2 will warn. The warning will also be emitted
during boot:

    $ erl -config dupkeys.config -s erlang halt
    2019-02-27 11:16:02.653100 application: ~ts; duplicate parameter: ~ts~n
        kernel
        key1
    =WARNING REPORT==== 27-Feb-2019::11:16:02.653100 ===
    application: kernel; duplicate parameter: key1

    $ erl -config dupapps.config -s erlang halt
    2019-02-27 11:16:02.653100 duplicate application config: ~ts
        kernel
    =WARNING REPORT==== 27-Feb-2019::11:16:02.653100 ===
    duplicate application config: kernel

Prior to this patch, the behaviour was unspecified,
and duplicate keys and duplicate apps would behave
different depending on the amount of config, the name
of the config files, and how those configs would be
listed. The goal is to raise an error in the future.