Add lib/compiler/scripts/smoke
Add `lib/compiler/scripts/smoke` for smoke testing the compiler (that
is, test that the compiler does not crash during compilation).

`smoke` first installs Elixir and mix. It then uses `mix` to download
a number of `hex` packages and compile them.

We don't intend to use `smoke` in our daily builds or Travis, but to run
it manually during compiler development.